ESLint bug fixes

+++ controller/post.js
@@ -1,13 +1,9 @@
-/* eslint-disable object-curly-spacing */
-/* eslint-disable indent */
-/* eslint-disable max-len */
-
 const knex = require('knex')(require('../knexfile.js').development);
-
 
 // controller to getAll Posts
 exports.getAll = (req, res) => {
-  // Select post and user fields by using a join between posts and users tables and order them chronologically, newest first
+  // Select post and user fields by using a join between posts and users tables
+  // and order them chronologically, newest first
   knex
     .select(
       'posts.id as post_id',
@@ -31,14 +27,12 @@ exports.getAll = (req, res) => {
       let updatedPosts = posts;
 
       // Check if user is logged in and update all logged in user's posts with "isCurrentUser" field
-      req.user && (
-        updatedPosts = updatedPosts.map((post) => {
-          return {
-            ...post,
-            isCurrentUser: post.users_id === req.user.id,
-          };
-        })
-      );
+      if (req.user) {
+        updatedPosts = updatedPosts.map((post) => ({
+          ...post,
+          isCurrentUser: post.users_id === req.user.id,
+        }));
+      }
 
       res.status(200).json(updatedPosts);
     })
@@ -74,14 +68,12 @@ exports.getOne = (req, res) => {
       let updatedPost = post;
 
       // Check if user is logged in and update all logged in user's posts with "isCurrentUser" field
-      req.user && (
-        updatedPost = updatedPost.map((post) => {
-          return {
-            ...post,
-            isCurrentUser: post.users_id === req.user.id,
-          };
-        })
-      );
+      if (req.user) {
+        updatedPost = updatedPost.map((post) => ({
+          ...post,
+          isCurrentUser: post.users_id === req.user.id,
+        }));
+      }
 
       res.status(200).json(...updatedPost);
     })
@@ -90,26 +82,28 @@ exports.getOne = (req, res) => {
     });
 };
 
-
 // controller to create a new Post
 exports.addOne = (req, res) => {
   // Deconstructing req.body for easier code digestion
-  const { title, description, category, offer, pic_url: picUrl } = req.body;
+  const {
+    title, description, category, offer, pic_url: picUrl,
+  } = req.body;
+
   // If user is not logged in, we don't allow them to create a new post
-  return (req.user === undefined) ? (
-    res.status(401).json({ message: 'Unauthorized' })
+  if (req.user === undefined) {
+    res.status(401).json({ message: 'Unauthorized' });
     // Validate request body for required fields
-  ) : (!title || !description || !category) ? (
-    res.status(400).json({ message: 'Missing post title or content fields' })
-  ) : (
+  } else if (!title || !description || !category) {
+    res.status(400).json({ message: 'Missing post title or content fields' });
+  } else {
     // Insert new post into DB: user_id comes from session, title and content from a request body
     knex('posts')
       .insert({
         user_id: req.user.id,
-        title: title,
-        description: description,
-        category: category,
-        offer: offer,
+        title,
+        description,
+        category,
+        offer,
         pic_url: picUrl,
         active: true,
       })
@@ -119,42 +113,42 @@ exports.addOne = (req, res) => {
       })
       .catch(() => {
         res.status(500).json({ message: 'Error creating a new post' });
-      })
-  );
+      });
+  }
 };
 
 // todo Test route
 exports.editOne = (req, res) => {
-  return (req.user === undefined) ? (
-    res.status(401).json({ message: 'Unauthorized' })
+  if (req.user === undefined) {
+    res.status(401).json({ message: 'Unauthorized' });
     // Validate request body for required fields
-  ) : (
+  } else {
     knex('posts')
       .update(req.body)
-      .where({ user_id: req.user.id, id: req.body.id })
+      .where({ user_id: req.user.id, id: req.params.postId })
       .then(() => {
-        res.send(`post with id: ${req.body.id} belonging to ${req.user.id} has been updated`);
+        res.send(`post with id: ${req.params.postId} belonging to ${req.user.id} has been updated`);
       })
       .catch((err) => {
-        res.send(`Error updating post with id: ${req.body.id} belonging to ${req.user.id} ${err}`).status(400);
-      })
-  );
+        res.send(`Error updating post with id: ${req.params.postId} belonging to ${req.user.id} ${err}`).status(400);
+      });
+  }
 };
 
 // todo Test route
 exports.deleteOne = (req, res) => {
-  return (req.user === undefined) ? (
-    res.status(401).json({ message: 'Unauthorized' })
+  if (req.user === undefined) {
+    res.status(401).json({ message: 'Unauthorized' });
+  } else {
     // Validate request body for required fields
-  ) : (
     knex('posts')
       .delete()
-      .where({ user_id: req.user.id, id: req.body.id })
+      .where({ user_id: req.user.id, id: req.params.postId })
       .then(() => {
-        res.status(204).send(`post with id: ${req.body.id} belonging to ${req.user.id} has been deleted`);
+        res.status(204).send(`post with id: ${req.params.postId} belonging to ${req.user.id} has been deleted`);
       })
       .catch((err) => {
-        res.status(400).send(`Error deleting post with id: ${req.body.id} belonging to ${req.user.id} ${err}`);
-      })
-  );
+        res.status(400).send(`Error deleting post with id: ${req.params.postId} belonging to ${req.user.id} ${err}`);
+      });
+  }
 };
